jdk1.8 替换成jdk17

+++ laokou-service/laokou-admin/laokou-admin-server/src/main/java/org/laokou/admin/server/application/service/SysMessageApplicationService.java
@@ -42,6 +42,6 @@ public interface SysMessageApplicationService {
 
     IPage<SysMessageVO> getUnReadList(HttpServletRequest request, SysMessageQO qo);
 
-    Integer unReadCount(HttpServletRequest request);
+    Long unReadCount(HttpServletRequest request);
 
 }
