+++ laokou-service/laokou-admin/laokou-admin-server/src/main/java/org/laokou/admin/server/application/service/impl/SysRoleApplicationServiceImpl.java
@@ -78,7 +78,7 @@ public class SysRoleApplicationServiceImpl implements SysRoleApplicationService 
     @Override
     public Boolean insertRole(SysRoleDTO dto, HttpServletRequest request) {
         SysRoleDO roleDO = ConvertUtil.sourceToTarget(dto, SysRoleDO.class);
-        int count = sysRoleService.count(Wrappers.lambdaQuery(SysRoleDO.class).eq(SysRoleDO::getName, roleDO.getName()).eq(SysRoleDO::getDelFlag, Constant.NO));
+        long count = sysRoleService.count(Wrappers.lambdaQuery(SysRoleDO.class).eq(SysRoleDO::getName, roleDO.getName()).eq(SysRoleDO::getDelFlag, Constant.NO));
         if (count > 0) {
             throw new CustomException("角色已存在，请重新填写");
         }
@@ -118,7 +118,7 @@ public class SysRoleApplicationServiceImpl implements SysRoleApplicationService 
     @Override
     public Boolean updateRole(SysRoleDTO dto, HttpServletRequest request) {
         SysRoleDO roleDO = ConvertUtil.sourceToTarget(dto, SysRoleDO.class);
-        int count = sysRoleService.count(Wrappers.lambdaQuery(SysRoleDO.class).eq(SysRoleDO::getName, roleDO.getName()).eq(SysRoleDO::getDelFlag, Constant.NO).ne(SysRoleDO::getId,roleDO.getId()));
+        long count = sysRoleService.count(Wrappers.lambdaQuery(SysRoleDO.class).eq(SysRoleDO::getName, roleDO.getName()).eq(SysRoleDO::getDelFlag, Constant.NO).ne(SysRoleDO::getId,roleDO.getId()));
         if (count > 0) {
             throw new CustomException("角色已存在，请重新填写");
         }
